+++ tests/testthat/test-method-dispatch.R
@@ -144,7 +144,7 @@ test_that("single dispatch fails with informative messages", {
     fail(Foo(x = 1))
   })
 
-  expect_error(fail(TRUE), class = "methodNotFound")
+  expect_error(fail(TRUE), class = "S7_error_method_not_found")
 })
 
 test_that("multiple dispatch fails with informative messages", {
@@ -160,7 +160,7 @@ test_that("multiple dispatch fails with informative messages", {
     fail(TRUE, TRUE)
   })
 
-  expect_error(fail(TRUE, TRUE), class = "methodNotFound")
+  expect_error(fail(TRUE, TRUE), class = "S7_error_method_not_found")
 })